Fix bug where particle text animation gets faster on every resize
The issue was that an additional animate call was being stacked on top
of an already existing recursion of animate calls. This meant the
animation would move 2 frames instead of 1 on every resize, and keep
increasing by one on each subsequent window resize. Cancel the previous
animation frame on resize to fix this.