Fix bug where home link doesn't work after scrolling down
On scroll, the intersection observer updates the active item in the
header to the next sections. However, the website link remains the same.
Since we are already at the home page link, clicking on "home" in the
header has no effect. Fix by adding a hash to the header nav link even
for the home page.